requirements.txt
google-cloud-pubsub==2.13.6
tweepy==4.10.1